Todo RESTful API - A TDD Approach to Building an API using NoSQL Database
Introduction
Testing is an integral part of the software development process, which helps
improve the quality of
the software. There are many types of testing involved
like manual testing, integration testing, functional testing, load testing, unit
testing, and other. In this project, we'll write our code following the rules of Test
Driven Development (TDD)


The REST architecture was originally designed to fit the
HTTP protocol
that the
world wide web uses.
Central to the concept of RESTful web services is the notion of resources.
Resources are represented by
URIs
. The clients send requests to these URIs
using
the methods defined by the HTTP protocol, and possibly as a result of that the
state of the affected resource changes.
The REST design
does not require a specific format for the data provided with
the requests. In general data is provided in the request body as a
JSON
blob, or
sometimes as arguments in the
query string
portion of the URL
Designing a simple
to
do
web service
The task of designing a web service or API that adheres to the REST guidelines
then becomes an exercise in identifying the resources that will be exposed and
ho
w they will be affected by the different request methods.
Let's say we want to write a To Do List application and we want to design a web
service for it. The first thing to do is to decide what is the root URL to access this
service. For example, we could
expose this service as:
http://[hostname]/todo/api/v1.0/
Here I have decided to include the name of the application and the version of the
API in the URL. Including the application name in the URL is usef
ul to provide a
namespace that separates this service from others that can be running on the
same system. Including the version in the URL can help with making updates in
the future, since new and potentially incompatible functions can be added under
a new
version, without affecting applications that rely on the older functions.
The next step is to select the resources that will be exposed by this service. This
is an extremely simple application, we only have tasks, so our only resource will
be the tasks in
our to do list.
Our tasks resource will use HTTP methods as follows:
HTTP
Method
URI
Action
We can define a task as having the following
fields:

id
: unique identifier for tasks. Numeric type.

title
: short task description. String type.

description
: long task description. Text type.

done
: task completion state. Boolean type

And with this we are basically done with the design part of our we
b service. All
that is left is to implement it!